requirements.txt
Appium-Python-Client==0.28
atomicwrites==1.2.1
attrs==18.2.0
bashplotlib==0.6.5
behave==1.2.6
certifi==2018.8.24
chardet==3.0.4
click==6.7
coloredlogs==10.0
cssselect==1.0.3
cycler==0.10.0
dataclasses==0.6
emoji==0.5.0
extras==1.0.0
first==2.0.1
fixtures==3.0.0
fuzzywuzzy==0.17.0
geographiclib==1.49
geopy==1.16.0
glob2==0.6
howdoi==1.1.13
humanfriendly==4.16.1
idna==2.7
jedi==0.12.1
kiwisolver==1.0.1
lettuce==0.2.23
lettuce-webdriver==0.3.5
linecache2==1.0.0
lxml==4.2.5
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.2.3
mock==2.0.0
more-itertools==4.3.0
nose==1.3.7
numpy==1.15.1
parse==1.8.4
parse-type==0.4.2
parso==0.3.1
pbr==4.2.0
Pillow==5.2.0
pip-tools==2.0.2
pipdeptree==0.13.0
pluggy==0.7.1
psycopg2==2.7.5
py==1.6.0
Pygments==2.2.0
PyHamcrest==1.9.0
pyparsing==2.2.0
pyquery==1.4.0
pytest==3.7.4
pytest-bdd==2.21.0
python-dateutil==2.7.3
python-mimeparse==1.6.0
python-subunit==1.3.0
pytz==2018.5
requests==2.19.1
requests-cache==0.4.13
ruamel.yaml==0.15.64
selenium==3.14.0
sh==1.12.14
six==1.11.0
sure==1.4.11
testtools==2.3.0
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.23
